conforms c99

+++ src/log.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
@@ -29,7 +30,7 @@ const char* drpc_now(char* buf, size_t bufsz) {
     size_t len = strftime(buf, bufsz, "%Y-%m-%d %H:%M:%S",
             localtime_r(&tv.tv_sec, &ltm));
     if (len <= bufsz) {
-        snprintf(buf + len, bufsz - len, ".%06ld", tv.tv_usec);
+        snprintf(buf + len, bufsz - len, ".%06d", (int32_t)tv.tv_usec);
     }
     return buf;
 }
